% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_Kendrick.R
\name{int_Kendrick}
\alias{int_Kendrick}
\title{Interactive Kendrick plot using plotly}
\usage{
int_Kendrick(data)
}
\arguments{
\item{data}{input the data created by the function read_ExcelResult}
}
\value{

}
\description{
Interactive Kendrick plot using plotly
}
\author{
Ghanimah Abuhaimed, Edwin Guevara Romero, Mohammad Hasyim Taufik
}
