% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_spectrum.R
\name{int_spectrum}
\alias{int_spectrum}
\title{Plot_ly version of the mass spectrum}
\usage{
int_spectrum(expMpZ, totAbund, Class)
}
\arguments{
\item{expMpZ}{input the data exponential of m/z ratio}

\item{totAbund}{input the data of total abundance}

\item{Class}{input the data Class}
}
\value{

}
\description{
Plot_ly version of the mass spectrum
}
\author{
Ghanimah Abuhaimed, Edwin Guevara Romero, Mohammad Hasyim Taufik
}
