% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kendrick_plot.R
\name{Kendrick_plot}
\alias{Kendrick_plot}
\title{Kendrick Plot}
\usage{
Kendrick_plot(data, main)
}
\arguments{
\item{data}{input the data created by the function read_ExcelResult}

\item{main}{input a custom name for the graph}
}
\value{

}
\description{
Kendrick Plot
}
\author{
Ghanimah Abuhaimed, Edwin Guevara Romero, Mohammad Hasyim Taufik
}
