% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_calculation.R
\name{load_calculation}
\alias{load_calculation}
\title{An Excel with Multiple Sheets Reader}
\usage{
load_calculation(fileDir)
}
\arguments{
\item{fileDir}{the relative path directory}
}
\value{
data framelib
}
\description{
This allows to simultaneously load all the sheets into a single data frame.
}
\examples{
load_calculation("../../final_project/example.xlsx")
}
\author{
Ghanimah Abuhaimed, Edwin Guevara Romero, Mohammad Hasyim Taufik
}
