% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compound.R
\name{plot_compound}
\alias{plot_compound}
\title{A Compound Class Plot Function}
\usage{
plot_compound(x, h, main, xlab, ylab)
}
\arguments{
\item{x}{data for the x axis}

\item{h}{data for the y axis}

\item{main}{title of the plot}

\item{xlab}{label of the x axis}

\item{ylab}{label of the y axis}
}
\description{
This allows you to plot a compound class plot given some inputs.
}
\examples{
plot_compound(x=rnorm(100),h=rnorm(100),"July","Abundance","Compound Class")
}
\author{
Ghanimah Abuhaimed, Edwin Guevara Romero, Mohammad Hasyim Taufik
}
\keyword{Kendrick}
\keyword{mass,}
\keyword{spectrometry}
