% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dbe.R
\name{plot_dbe}
\alias{plot_dbe}
\title{A Double Bond Equivalent Scatter Plot Function}
\usage{
plot_dbe(colorPalette, abundance, n, x, y, size, main, pch, xlab, ylab)
}
\arguments{
\item{colorPalette}{list of the color palette}

\item{abundance}{intensity data of the spectometry data}

\item{n}{number of data for segmenting the continuous color palette}

\item{x}{data for the x axis}

\item{y}{data for the y axis}

\item{size}{size of the point relative to the abundance data}

\item{main}{title of the plot}

\item{pch}{type of point}

\item{xlab}{label of the x axis}

\item{ylab}{label of the y axis}
}
\description{
This allows you to plot a double bond equivalent scatter plot given some inputs.
}
\examples{
plot_dbe(colorPalette=c('blue','green','yellow','red'),
abundance=rnorm(100),100,x=rnorm(100),y=rnorm(100),0.00000001,"Kendrick Plot",16,"x","y")
}
\author{
Ghanimah Abuhaimed, Edwin Guevara Romero, Mohammad Hasyim Taufik
}
\keyword{Kendrick}
\keyword{mass,}
\keyword{spectrometry}
