% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DBE_Het.R
\name{plot_DBE_Het}
\alias{plot_DBE_Het}
\title{Plot a scatter plot of the DBE for each element}
\usage{
plot_DBE_Het(data, Het, main)
}
\arguments{
\item{data}{input the data created by the function read_ExcelResult}

\item{Het}{input heteroatom as #C or #H or #N or #O or #S}

\item{main}{input a custom name for the graph}
}
\value{
A scatter plot of the DBE for each element, gives information about its abundance by the size of the spot and its color
}
\description{
Plot a scatter plot of the DBE for each element
}
\author{
Ghanimah Abuhaimed, Edwin Guevara Romero, Mohammad Hasyim Taufik
}
