% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DBE_Hist.R
\name{plot_DBE_Hist}
\alias{plot_DBE_Hist}
\title{Histogram of DBE by class compound}
\usage{
plot_DBE_Hist(data, main, col, classcmp)
}
\arguments{
\item{data}{input the data created by the function read_ExcelResult}

\item{main}{input a custom name for the graph}

\item{col}{input a custom color for the graph}

\item{classcmp}{Class compound which DBE is plotted.}
}
\value{

}
\description{
Histogram of DBE by class compound
}
\author{
Ghanimah Abuhaimed, Edwin Guevara Romero, Mohammad Hasyim Taufik
}
