% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_massspectrum.R
\name{plot_massspectrum}
\alias{plot_massspectrum}
\title{Plot the mass spectrum}
\usage{
plot_massspectrum(data, main, col, ylim, xlim)
}
\arguments{
\item{data}{input the data created by the function read_ExcelResult}

\item{main}{input a custom name for the graph}

\item{col}{input a custom color for the graph}

\item{ylim}{input the limits for y-axis}

\item{xlim}{input the limits for x_axis}
}
\value{
A mass spectrum plot
}
\description{
Plot the mass spectrum
}
\author{
Ghanimah Abuhaimed, Edwin Guevara Romero, Mohammad Hasyim Taufik
}
