% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spectrum.R
\name{plot_spectrum}
\alias{plot_spectrum}
\title{A Mass Spectrum Plot Function}
\usage{
plot_spectrum(color, x, y, main, pch, xlab, ylab)
}
\arguments{
\item{color}{list of the color palette}

\item{x}{data for the x axis}

\item{y}{data for the y axis}

\item{main}{title of the plot}

\item{pch}{type of point}

\item{xlab}{label of the x axis}

\item{ylab}{label of the y axis}
}
\description{
This allows you to plot a mass spectrum plot given some inputs.
}
\examples{
plot_spectrum(x=rnorm(100), y=rnorm(100), pch = 10, col = "blue",
main = "Mass Spectra Plot", ylab = "abundance", xlab = "m/z")
}
\author{
Ghanimah Abuhaimed, Edwin Guevara Romero, Mohammad Hasyim Taufik
}
\keyword{Kendrick}
\keyword{mass,}
\keyword{spectrometry}
