% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ExcelResult.R
\name{read_ExcelResult}
\alias{read_ExcelResult}
\title{Load the Excel files given as result of a FTICRMS analysis by CoreLabs.}
\usage{
read_ExcelResult(data)
}
\arguments{
\item{data}{Input the name of the Excel file with .xlsx extension.}
}
\value{
a list named as the excel file with the data organize according to the names of the sheets
}
\description{
Load the Excel files given as result of a FTICRMS analysis by CoreLabs.
}
\author{
Ghanimah Abuhaimed, Edwin Guevara Romero, Mohammad Hasyim Taufik
}
