% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Van_Krevelenplot.R
\name{Van_Krevelenplot}
\alias{Van_Krevelenplot}
\title{Van Krevelen plot}
\usage{
Van_Krevelenplot(data, main, ratio, xlab, ylab, xlim, ylim)
}
\arguments{
\item{data}{input the data created by the function read_ExcelResult}

\item{main}{input a custom name for the graph}

\item{ratio}{ratio of elements to plot as are designed in the excel file: H/C, N/C, O/C or S/C}

\item{xlab}{name of the x-axis}

\item{ylab}{name of the y-axis}

\item{xlim}{limits for x-axis}

\item{ylim}{limits for y-axis}
}
\value{

}
\description{
Van Krevelen plot
}
\author{
Ghanimah Abuhaimed, Edwin Guevara Romero, Mohammad Hasyim Taufik
}
